\docType{package}
\name{blat-package}
\alias{blat}
\alias{blat-package}
\title{Methods to import, export and manipulate BLAT result data.}
\description{
  This pacakge enables the import, export, and analysis of
  BLAT result files, mostly in the form of PSL-formatted
  BLAT alignment results. This code is really old, so the
  package is a work in progress. IIRC, I ported some of Jim
  Kent's orgiinal C code to calculate alignment scores
  etc...
}
\author{
  Mark Cowley <m.cowley@garvan.org.au>
}
\seealso{
  \code{\link{import.psl}}, \code{\link{pslScore}},
  \code{\link{write.psl.track}}
}

